refactor(cdk): Split CDK stack definitions across separate files
Building on https://github.com/guardian/cdk-playground/pull/1058,
in an attempt to further improve maintainability,
split the EC2, Lambda and ECS stacks into separate files.
