+++ cdk/bin/cdk.ts
@@ -1,10 +1,8 @@
 import 'source-map-support/register';
 import { GuRoot } from '@guardian/cdk/lib/constructs/root';
-import {
-	CdkPlaygroundEc2,
-	CdkPlaygroundEcs,
-	CdkPlaygroundLambda,
-} from '../lib/cdk-playground';
+import { CdkPlaygroundEc2 } from '../lib/cdk-playground-ec2';
+import { CdkPlaygroundEcs } from '../lib/cdk-playground-ecs';
+import { CdkPlaygroundLambda } from '../lib/cdk-playground-lambda';
 import { EventForwarder } from '../lib/event-forwarder';
 
 const app = new GuRoot();
@@ -18,7 +16,6 @@ const ec2Stack = new CdkPlaygroundEc2(app, 'CdkPlaygroundEc2-CODE', {
 
 new CdkPlaygroundLambda(app, 'CdkPlaygroundLambda-CODE', {
 	cloudFormationStackName: 'deploy-CODE-cdk-playground-lambda',
-	buildIdentifier: process.env.GITHUB_RUN_NUMBER ?? 'DEV',
 });
 
 new CdkPlaygroundEcs(app, 'CdkPlaygroundEcs-CODE', {
