+++ cdk/lib/cdk-playground-ecs.ts
@@ -1,8 +1,9 @@
-import { GuApiLambda } from '@guardian/cdk';
-import { AccessScope } from '@guardian/cdk/lib/constants/access';
 import { GuCertificate } from '@guardian/cdk/lib/constructs/acm';
-import type { GuStackProps } from '@guardian/cdk/lib/constructs/core';
-import { GuParameter, GuStack } from '@guardian/cdk/lib/constructs/core';
+import {
+	GuParameter,
+	GuStack,
+	type GuStackProps,
+} from '@guardian/cdk/lib/constructs/core';
 import { GuCname } from '@guardian/cdk/lib/constructs/dns';
 import {
 	GuHttpsEgressSecurityGroup,
@@ -15,15 +16,9 @@ import {
 	GuApplicationTargetGroup,
 	GuHttpsApplicationListener,
 } from '@guardian/cdk/lib/constructs/loadbalancing';
-import { GuEc2AppExperimental } from '@guardian/cdk/lib/experimental/patterns/ec2-app';
 import type { App } from 'aws-cdk-lib';
 import { Duration } from 'aws-cdk-lib';
-import {
-	InstanceClass,
-	InstanceSize,
-	InstanceType,
-	Vpc,
-} from 'aws-cdk-lib/aws-ec2';
+import { Vpc } from 'aws-cdk-lib/aws-ec2';
 import { Repository } from 'aws-cdk-lib/aws-ecr';
 import {
 	Cluster,
@@ -33,9 +28,8 @@ import {
 	LogDriver,
 	VersionConsistency,
 } from 'aws-cdk-lib/aws-ecs';
-import { Runtime } from 'aws-cdk-lib/aws-lambda';
 
-interface CdkPlaygroundProps extends Omit<GuStackProps, 'stack' | 'stage'> {
+interface CdkPlaygroundEcsProps extends Omit<GuStackProps, 'stack' | 'stage'> {
 	/**
 	 * Which application build to run.
 	 * This will typically match the build number provided by CI.
@@ -46,105 +40,10 @@ interface CdkPlaygroundProps extends Omit<GuStackProps, 'stack' | 'stage'> {
 	buildIdentifier: string;
 }
 
-export class CdkPlaygroundEc2 extends GuStack {
-	constructor(scope: App, id: string, props: CdkPlaygroundProps) {
-		super(scope, id, {
-			...props,
-			stack: 'deploy',
-			stage: 'CODE',
-			env: { region: 'eu-west-1' },
-		});
-
-		const { buildIdentifier } = props;
-
-		const ec2AppDomainName = 'cdk-playground.code.dev-gutools.co.uk';
-
-		const ec2App = 'cdk-playground';
-
-		const { loadBalancer } = new GuEc2AppExperimental(this, {
-			buildIdentifier,
-			applicationPort: 9000,
-			app: ec2App,
-			instanceType: InstanceType.of(InstanceClass.T4G, InstanceSize.MICRO),
-			access: { scope: AccessScope.PUBLIC },
-			userData: {
-				distributable: {
-					fileName: `${ec2App}-${buildIdentifier}.deb`,
-					executionStatement: `dpkg -i /${ec2App}/${ec2App}-${buildIdentifier}.deb`,
-				},
-			},
-			certificateProps: {
-				domainName: ec2AppDomainName,
-			},
-			monitoringConfiguration: { noMonitoring: true },
-			scaling: {
-				minimumInstances: 1,
-				maximumInstances: 10,
-			},
-			applicationLogging: {
-				enabled: true,
-				systemdUnitName: 'cdk-playground',
-			},
-			imageRecipe: 'arm64-jammy-java21-deploy-infrastructure',
-			instanceMetricGranularity: '5Minute',
-		});
-
-		new GuCname(this, 'EC2AppDNS', {
-			app: ec2App,
-			ttl: Duration.hours(1),
-			domainName: ec2AppDomainName,
-			resourceRecord: loadBalancer.loadBalancerDnsName,
-		});
-	}
-}
-
-export class CdkPlaygroundLambda extends GuStack {
-	constructor(scope: App, id: string, props: CdkPlaygroundProps) {
-		super(scope, id, {
-			...props,
-			stack: 'deploy',
-			stage: 'CODE',
-			env: { region: 'eu-west-1' },
-		});
-
-		const lambdaDomainName = 'cdk-playground-lambda.code.dev-gutools.co.uk';
-		const lambdaApp = 'cdk-playground-lambda';
-
-		const lambda = new GuApiLambda(this, 'lambda', {
-			fileName: `cdk-playground-lambda.zip`,
-			handler: 'handler.main',
-			runtime: Runtime.NODEJS_24_X,
-			monitoringConfiguration: {
-				noMonitoring: true,
-			},
-			app: lambdaApp,
-			api: {
-				id: `${lambdaApp}-api`,
-				description: lambdaApp,
-			},
-		});
-
-		const domain = lambda.api.addDomainName('domain', {
-			domainName: lambdaDomainName,
-			certificate: new GuCertificate(this, {
-				app: lambdaApp,
-				domainName: lambdaDomainName,
-			}),
-		});
-
-		new GuCname(this, 'LambdaDNS', {
-			app: lambdaApp,
-			ttl: Duration.hours(1),
-			domainName: lambdaDomainName,
-			resourceRecord: domain.domainNameAliasDomainName,
-		});
-	}
-}
-
 // For now, we provision all of this infrastructure via constructs as part of this repo.
 // Once we've figured out the details we'd aim to provide this to users via a GuCDK pattern instead.
 export class CdkPlaygroundEcs extends GuStack {
-	constructor(scope: App, id: string, props: CdkPlaygroundProps) {
+	constructor(scope: App, id: string, props: CdkPlaygroundEcsProps) {
 		super(scope, id, {
 			...props,
 			stack: 'deploy',
