+++ cdk/lib/cdk-playground-lambda.ts
@@ -1,0 +1,52 @@
+import { GuApiLambda } from '@guardian/cdk';
+import { GuCertificate } from '@guardian/cdk/lib/constructs/acm';
+import { GuStack, type GuStackProps } from '@guardian/cdk/lib/constructs/core';
+import { GuCname } from '@guardian/cdk/lib/constructs/dns';
+import type { App } from 'aws-cdk-lib';
+import { Duration } from 'aws-cdk-lib';
+import { Runtime } from 'aws-cdk-lib/aws-lambda';
+
+type CdkPlaygroundLambdaProps = Omit<GuStackProps, 'stack' | 'stage'>;
+
+export class CdkPlaygroundLambda extends GuStack {
+	constructor(scope: App, id: string, props: CdkPlaygroundLambdaProps) {
+		super(scope, id, {
+			...props,
+			stack: 'deploy',
+			stage: 'CODE',
+			env: { region: 'eu-west-1' },
+		});
+
+		const lambdaDomainName = 'cdk-playground-lambda.code.dev-gutools.co.uk';
+		const lambdaApp = 'cdk-playground-lambda';
+
+		const lambda = new GuApiLambda(this, 'lambda', {
+			fileName: `cdk-playground-lambda.zip`,
+			handler: 'handler.main',
+			runtime: Runtime.NODEJS_24_X,
+			monitoringConfiguration: {
+				noMonitoring: true,
+			},
+			app: lambdaApp,
+			api: {
+				id: `${lambdaApp}-api`,
+				description: lambdaApp,
+			},
+		});
+
+		const domain = lambda.api.addDomainName('domain', {
+			domainName: lambdaDomainName,
+			certificate: new GuCertificate(this, {
+				app: lambdaApp,
+				domainName: lambdaDomainName,
+			}),
+		});
+
+		new GuCname(this, 'LambdaDNS', {
+			app: lambdaApp,
+			ttl: Duration.hours(1),
+			domainName: lambdaDomainName,
+			resourceRecord: domain.domainNameAliasDomainName,
+		});
+	}
+}
